[Refactor] Reuse Diffusers LTX-2.5 decoder
